feat: implementando validações e tratamento de erros e configuração para o relatório

+++ src/main/java/com/br/gestaoPedidos/produtos/service/ProdutoService.java
@@ -2,10 +2,12 @@ package com.br.gestaoPedidos.produtos.service;
 
 import java.time.LocalDateTime;
 import java.util.List;
-import java.util.UUID;
 
 import org.springframework.stereotype.Service;
 
+import com.br.gestaoPedidos.pedidos.repository.PedidoProdutoRepository;
+import com.br.gestaoPedidos.produtos.exception.ProdutoJaExisteException;
+import com.br.gestaoPedidos.produtos.exception.ProdutoVinculadoPedidoException;
 import com.br.gestaoPedidos.produtos.model.ProdutoModel;
 import com.br.gestaoPedidos.produtos.repository.ProdutoRepository;
 
@@ -13,30 +15,38 @@ import com.br.gestaoPedidos.produtos.repository.ProdutoRepository;
 public class ProdutoService {
 
 	private ProdutoRepository produtoRepository;
+	private PedidoProdutoRepository pedidoProdutoRepository;
 
-	public ProdutoService(ProdutoRepository produtoRepository) {
+	public ProdutoService(ProdutoRepository produtoRepository, PedidoProdutoRepository pedidoProdutoRepository) {
 		this.produtoRepository = produtoRepository;
+		this.pedidoProdutoRepository = pedidoProdutoRepository;
 	}
 
 	public List<ProdutoModel> listar() {
 		return produtoRepository.findAll();
 	}
 
-	public ProdutoModel buscarPorId(UUID id) {
-		return produtoRepository.findById(id).orElseThrow(() -> new RuntimeException("Produto não encontrado"));
+	public ProdutoModel buscarPorId(Long id) {
+		return produtoRepository.findById(id).orElseThrow(() -> new ProdutoJaExisteException("produto.naoEncontrado"));
 
 	}
 
 	public ProdutoModel criar(ProdutoModel produto) {
+		boolean existeProduto = produtoRepository.existsByDescricao(produto.getDescricao());
+
+		if (existeProduto) {
+			throw new ProdutoJaExisteException("produto.jaExistente");
+		}
+
 		produto.setDataCriacao(LocalDateTime.now());
 
 		return produtoRepository.save(produto);
 
 	}
 
-	public ProdutoModel alterar(UUID id, ProdutoModel produto) {
+	public ProdutoModel alterar(Long id, ProdutoModel produto) {
 		ProdutoModel produtoExistente = produtoRepository.findById(id)
-				.orElseThrow(() -> new RuntimeException("Pedido não encontrado"));
+				.orElseThrow(() -> new ProdutoJaExisteException("produto.naoEncontrado"));
 
 		produto.setId(id);
 		produto.setDataCriacao(produtoExistente.getDataCriacao());
@@ -45,7 +55,12 @@ public class ProdutoService {
 		return produtoRepository.save(produto);
 	}
 
-	public void deletar(UUID id) {
+	public void deletar(Long id) {
+
+		if (!this.pedidoProdutoRepository.findByProdutoId(id).isEmpty()) {
+			throw new ProdutoVinculadoPedidoException("produto.vinculadoPedido");
+		}
+
 		produtoRepository.deleteById(id);
 	}
